Add utility methods for angle conversion
