+++ ColorMine.Test/ColorSpaces/Comparisons/Cie1976ComparisonTest.cs
@@ -1,6 +1,6 @@
 ﻿using ColorMine.ColorSpaces;
 using ColorMine.ColorSpaces.Comparisons;
-using ColorMine.ColorSpaces.Conversions.Utility;
+using ColorMine.ColorSpaces.Utility;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace ColorMine.Test.ColorSpaces.Comparisons
