+++ ColorMine.Test/Utility/DoubleExtensionTest.cs
@@ -1,4 +1,4 @@
-﻿using ColorMine.ColorSpaces.Conversions.Utility;
+﻿using ColorMine.ColorSpaces.Utility;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace ColorMine.Test.Utility
