+++ ColorMine/ColorSpaces/Conversions/CmykConverter.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using ColorMine.ColorSpaces.Conversions.Utility;
+﻿using ColorMine.ColorSpaces.Utility;
 
 namespace ColorMine.ColorSpaces.Conversions
 {
